Added happiness questions/categories seed data.

+++ db/migrate/20140506021248_create_happiness_questions.rb
@@ -2,7 +2,7 @@ class CreateHappinessQuestions < ActiveRecord::Migration
   def change
     create_table :happiness_questions do |t|
       t.text :content
-      t.float :max_score
+      t.integer :max_score
 
       t.timestamps
     end
