+++ db/migrate/20140506021401_create_happiness_responses.rb
@@ -2,7 +2,7 @@ class CreateHappinessResponses < ActiveRecord::Migration
   def change
     create_table :happiness_responses do |t|
       t.references :happiness_question, index: true
-      t.float :value
+      t.integer :value
       t.references :user, index: true
 
       t.timestamps
